Update README.md
# pcl_indoorPlaneSegmentation
This PCL perception tool is used to identify target plane for a semi autonomous painting robot. Potential planes within working distance will be identified, then further classifided to their type. The pointcloud will then be stored in a vector of struct for further used. A visualizer is used to visualized the segmentation result. 

### Example 1
![alt text](/documentations/example1.png?)


### Example 2
![alt text](/documentations/example2.png?)